Desktop/SpotAI/requirements.txt
flask==2.3.3
spotipy==2.23.0
openai==1.3.0
numpy==1.26.4
torch==2.1.0
transformers==4.36.0
scikit-learn==1.3.0
python-dotenv==1.0.0
flask-caching==2.1.0
